Update README.md
This project is an implementation using the ToDoList react library, an application that lists activities to be performed and marks those that have already been done. The user types the name of the activity through the text input field, insert it in the list with the insert button, to mark an activity as done, just check the check button in each one of them. In addition to showing the list of activities, this version of ToDoList also shows how many of the activities in the list are completed.